+++ lib/app/.temp/siteData.d.ts
@@ -1,3 +1,0 @@
-import { SiteData } from "../../prepare/resolveOptions";
-export const siteData: SiteData;
-export const PageComponents: { [key: string]: any };