+++ lib/app/components/NotFound.tsx
@@ -1,3 +1,0 @@
-import * as React from 'react';
-
-export default () => <div className="section">404 | Not Found</div>;